Rename doc module to docs
