+++ arbi/src/docs/algorithms.rs
