+++ arbi/src/lib.rs
@@ -33,7 +33,7 @@ mod division;
 #[cfg(not(doctest))]
 #[allow(unknown_lints)]
 #[allow(clippy::doc_lazy_continuation)]
-pub mod doc;
+pub mod docs;
 mod exponentiation;
 mod fits;
 mod floor;
